+++ packages/frontend/src/redux/selectors/crossStateSelectors/selectRecoveryLoader.js
@@ -1,7 +1,7 @@
 import { createSelector } from 'reselect';
 
-import { selectAccountId } from '../slices/account';
-import { selectRecoveryMethodsByAccountId, selectRecoveryMethodsLoading } from '../slices/recoveryMethods';
+import { selectAccountId } from '../../slices/account';
+import { selectRecoveryMethodsByAccountId, selectRecoveryMethodsLoading } from '../../slices/recoveryMethods';
 
 export default createSelector(
     [selectRecoveryMethodsLoading, selectAccountId, selectRecoveryMethodsByAccountId],
